requirements.txt
hydra-core==1.3.2          # 保持配置管理功能，升级到稳定新版本
opencv-python==4.9.0.80     #保持OpenCV功能，升级到最新稳定版
#torch==2.3.0+cu121         # 升级到支持CUDA 12.1的最新稳定版（向下兼容12.7）
#torchvision==0.18.0+cu121  # 对应torch 2.3.0的版本
omegaconf==2.3.0           # Hydra的依赖项升级到兼容版本
tqdm==4.66.4               # 升级到最新稳定版
pyquaternion==0.9.9        # 保持原版（无重大更新）
imgaug==0.4.0              # 保持原版（后续版本变化较大可能破坏兼容性）
torchmetrics==1.4.0        # 升级到适配PyTorch 2.x的版本
pytorch-lightning==2.2.4   # 升级到支持PyTorch 2.x的LTS版本
fvcore==0.1.5.post20221221 # 升级到最新维护版本
einops==0.8.0              # 升级到最新稳定版
efficientnet-pytorch==0.7.1 # 保持原版（测试无兼容性问题）
wandb==0.18.0              # 升级到最新稳定版
